[#93] Renames chatbot directory
